+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - On X11, drag-and-drop receiving an unsupported drop type can no longer cause the WM to freeze.
 - Fix issue whereby the OpenGL context would not appear at startup on macOS Mojave (#1069).
 - **Breaking:** Removed `From<NSApplicationActivationPolicy>` impl from `ActivationPolicy` on macOS.
+- On macOS, the application can request the user's attention with `WindowExt::request_user_attention`.
 
 # Version 0.17.2 (2018-08-19)